Login screen appears first now/FireBase working
